requirements.txt
streamlit==1.31.1
Pillow==10.1.0 